requirements.txt
bcrypt==4.2.0
black==24.10.0
click==8.1.7
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
psycopg==3.2.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
typing_extensions==4.12.2
